Initial import of 64-bit support for cuckoomon, completely untested so far.
Also protected lasterror during calls to get_ip_port, thanks to KillerInstinct for the report

Some notes about the port:
We've split the hooking code where necessary into 32bit and 64bit files -- the 64bit .c is nearly completely new.
x64 cuckoomon DLL is built as cuckoomon_x64.dll
PEB/TEB accesses have been abstracted
We provide binaries for x32 and x64 that can perform the appropriate injection -- Cuckoo itself will no longer perform injection, though I may leave in 32bit injection compatibility for those who haven't compiled the new loader binaries
g_hooks is no longer a huge array in the DLL, the hook-specific payloads have been separated out into per-hook allocations which on x64 will be located within 4GB of the function being hooked, this allows us to not have to craft equivalent forms of various instructions using rip-relative memory addressing.
I implemented a hackish change to capstone so we can find the offset of rip-relative displacements -- I honestly don't know why so many projects have switched to capstone (let alone cuckoomon as it previously was only using it as an LDE) except for jumping on bandwagons; it's really a pretty terrible engine for any uses other than printing disassembly.  The fact that capstone's own disassembly printer can't work off capstone's own public instruction interface is obvious evidence of this.
The 32bit version of cuckoomon should still operate exactly the same as before

The next changes will involve changes to Cuckoo and logging formats

+++ config.c
@@ -56,7 +56,7 @@ int read_config(void)
                     ARRAYSIZE(g_config.results));
             }
 			else if (!strcmp(key, "file-of-interest")) {
-				unsigned len = strlen(value);
+				unsigned int len = (unsigned int)strlen(value);
 				if (len > 1) {
 					if (value[1] == ':') {
 						// is a file
@@ -64,7 +64,7 @@ int read_config(void)
 						wchar_t *utmp = calloc(1, MAX_PATH * sizeof(wchar_t));
 						unsigned int full_len;
 						ensure_absolute_ascii_path(tmp, value);
-						full_len = strlen(tmp);
+						full_len = (unsigned int)strlen(tmp);
 						for (unsigned int i = 0; i < full_len; i++)
 							utmp[i] = (wchar_t)(unsigned short)tmp[i];
 						free(tmp);
@@ -77,7 +77,7 @@ int read_config(void)
 					else {
 						// is a URL
 						wchar_t *utmp = calloc(1, 512 * sizeof(wchar_t));
-						unsigned int url_len = strlen(value);
+						unsigned int url_len = (unsigned int)strlen(value);
 						for (unsigned int i = 0; i < url_len; i++)
 							utmp[i] = (wchar_t)(unsigned short)value[i];
 						g_config.url_of_interest = utmp;
