+++ utf8.c
@@ -48,7 +48,8 @@ int utf8_length(unsigned short x)
 
 int utf8_strlen_ascii(const char *s, int len)
 {
-    if(len < 0) len = strlen(s);
+    if(len < 0)
+		len = (int)strlen(s);
 
     int ret = 0;
     while (len-- != 0) {
@@ -70,7 +71,8 @@ int utf8_strlen_unicode(const wchar_t *s, int len)
 
 char * utf8_string(const char *str, int length)
 {
-    if (length == -1) length = strlen(str);
+	if (length == -1)
+		length = (int)strlen(str);
 
     int encoded_length = utf8_strlen_ascii(str, length);
     char * utf8string = (char *) malloc(encoded_length+4);
